Remove a TODO that isn't going to happen
In the Listenable approach, each Listener needs a parent struct holding
stuff captured from the stack. For one thing, this gets quite verbose and
ugly. But more problematically, there's no good place for the parent struct
to live. It can't go on the caller's stack, because it would disappear too
quickly. But it can't cleanly go in the Listenable either -- the Listenable
would have a heterogeneous collection of parent structs, and wouldn't know
how to deinit and dealloc them.