Add the meager beginnings of a Rust impl
The Rust learning curve is every bit as tough as reported. Model currently
has a field of type RefCell<Option<Rc<RefCell<T>>>>, which is ridiculous
... but nothing else I tried worked, and I tried rather a lot of things.

Would be nice to either find a better approach, or understand better why
this approach is necessary (e.g. why Box, Cell, &, and &mut all turned out
to be unhelpful).

Would also be nice to understand why assigning a RefMut to a local seems to
break everything.

And finally, would be nice to know why if-let destructuring doesn't work
with a Ref on the RHS, but does work if you do &*Ref. Is that idiomatic
and/or a reasonable thing to do?