Reorganize things to make ownership issues easier to see
This occasionally segfaults on exit. I believe what's happening is that the
main fn returns, and the SimControlImpl and all its contents get popped off
the stack, but the simulator thread is still running, and still trying to
access the SimControlImpl.

The fundamental challenge is communicating between the main thread and the
simulator thread. This challenge was present before; the code reorg just
makes it easier to see.