Move most of the simulation config to main
This required switching back to async/await. I don't see a good way to get
Thread.spawn() to run a fn that takes a parameterized arg type, and in this
case we need to pass in something like SimContext(N,P).

Still need to move the accelerators to main. This isn't completely trivial,
because the constructors need access to particle-state info that exists
only as stack variables inside runSimulation.